fix: support passing CAT in CloudFront querystring
The code for fetching a CAT from the querystring looks at the incoming
URL. The code for translating a Lambda@edge request to the expected
`node:http` `IncomingMessage` doesn't include the querystring at all,
leading to it being impossible to pass a CAT in the querystring when
using Lambda@edge.

This commit fixes this by including the querystring in the passed URL if
it is set.